+++ src/lib/createFakeConnection.ts
@@ -1,6 +1,7 @@
 import { stub } from 'sinon';
 import * as typeorm from 'typeorm';
 import { FakeConnectionManager } from './FakeConnectionManager';
+import { DEFAULT_CONNECTION_NAME } from '@nestjs/typeorm/dist/typeorm.constants';
 
 // The fake connection, here we can get the queries and parameters.
 export const fakeConnection = {
@@ -20,7 +21,9 @@ export const getFakeConnectionManager = () => {
 export const createFakeConnection = stub(typeorm, 'createConnection').callsFake(
   async function(optionsOrName: typeorm.ConnectionOptions) {
     const connectionName =
-      typeof optionsOrName === 'string' ? optionsOrName : 'default';
+      typeof optionsOrName === 'string'
+        ? optionsOrName
+        : DEFAULT_CONNECTION_NAME;
     const options =
       optionsOrName instanceof Object
         ? optionsOrName