+++ src/lib/typeorm-test.module.ts
@@ -3,19 +3,43 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { ConnectionOptions } from 'typeorm';
 
 import { TypeOrmTestCoreModule } from './typeorm-test-core.module';
+import { DEFAULT_CONNECTION_NAME } from '@nestjs/typeorm/dist/typeorm.constants';
+
+interface TypeOrmTestModuleOptions {
+  entities: Function[];
+  type?: string;
+  name?: string;
+}
+
+const TYPE = 'postgres';
 
 @Module({})
 export class TypeOrmTestModule extends TypeOrmModule {
-  static forTest(entities: Function[] = [], type = 'postgres'): DynamicModule {
+  static forTest(
+    entitiesOrOptions: TypeOrmTestModuleOptions | Function[],
+  ): DynamicModule {
+    let options;
+
+    if (!(entitiesOrOptions instanceof Array)) {
+      const _options = entitiesOrOptions as TypeOrmTestModuleOptions;
+      options = {
+        type: _options.type || TYPE,
+        name: _options.name || DEFAULT_CONNECTION_NAME,
+        entities: _options.entities || [],
+      };
+    } else {
+      entitiesOrOptions = entitiesOrOptions || [];
+      options = {
+        type: TYPE,
+        name: DEFAULT_CONNECTION_NAME,
+        entities: entitiesOrOptions,
+      };
+    }
+
     // The type serves as the handler to create the queries, it should be one of the
     // supported types of typeorm.
-    const options = {
-      type,
-      entities,
-    } as ConnectionOptions;
-
-    const root = TypeOrmTestCoreModule.forRoot(options);
-    const feature = this.forFeature(entities);
+    const root = TypeOrmTestCoreModule.forRoot(options as ConnectionOptions);
+    const feature = this.forFeature(options.entities, options.name);
 
     const result = {
       module: TypeOrmTestModule,