+++ src/tests/shared/User.ts
@@ -1,0 +1,19 @@
+import { Entity, BaseEntity, PrimaryGeneratedColumn, Column } from 'typeorm';
+
+@Entity()
+export class User extends BaseEntity {
+  @PrimaryGeneratedColumn()
+  id: number;
+
+  @Column()
+  name: string;
+
+  @Column('text', { unique: true })
+  username: string;
+
+  @Column('text', { unique: true })
+  email: string;
+
+  @Column()
+  password: string;
+}